requirements.txt
kafka_python==2.0.2
psycopg2-binary==2.9.9
apache-airflow==2.7.3
pyspark==3.5.0
unidecode==1.3.7
pandas==2.2.2